Modify move secure byte array to security namespace
